Fix warning about missing model-id in Ollama
Fixes: #1043
