+++ model-providers/ollama/runtime/src/main/java/io/quarkiverse/langchain4j/ollama/runtime/config/ModelIdConfigRelocateInterceptor.java
@@ -1,9 +1,29 @@
 package io.quarkiverse.langchain4j.ollama.runtime.config;
 
+import java.util.function.Function;
+
 import io.smallrye.config.RelocateConfigSourceInterceptor;
 
 public class ModelIdConfigRelocateInterceptor extends RelocateConfigSourceInterceptor {
+
+    private static final Function<String, String> RELOCATE = new Function<>() {
+        @Override
+        public String apply(String name) {
+            if (!name.startsWith("quarkus.langchain4j.ollama")) {
+                return name;
+            }
+            if (!name.endsWith("model-name")) {
+                return name;
+            }
+            int index = name.lastIndexOf(".model-name");
+            if (index < 1) {
+                return name;
+            }
+            return name.substring(0, index) + ".model-id";
+        }
+    };
+
     public ModelIdConfigRelocateInterceptor() {
-        super(ModelIdToModelNameFunction.INSTANCE);
+        super(RELOCATE);
     }
 }
